tests/webapi-w3c-fileapi-tests/fileapi/FileReader_abort_when_readyState_EMPTY.html

<!DOCTYPE html>
<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification, 
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list 
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice, 
  this list of conditions and the following disclaimer in the documentation 
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors 
  may be used to endorse or promote products derived from this work without 
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS" 
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE 
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE 
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT, 
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, 
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, 
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY 
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING 
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, 
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.  
  
Authors:
        Fan,Weiwei <weiwix.fan@intel.com>

-->

<html>
  <head>
    <title>File API Test: FileReader_abort_when_readyState_EMPTY</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="http://www.w3.org/TR/2011/WD-FileAPI-20111020/#dfn-filereader" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check if FileReader.abort is called when readyState is EMPTY that expecting FileReader.result is to be set null" />
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
  </head>
    <div id="log"></div>
    <script type="text/javascript">
        var t = async_test(document.title);
        try {
            var reader = new FileReader();
            // EMPTY is the default state of a newly minted FileReader object, until one of the read methods have been called on it
            if (reader.readyState == reader.EMPTY) {
                reader.abort();
                reader.onabort = t.step_func(function () {
                    // if the readyState is EMPTY (no read method has been called) then the result attribute MUST return null
                    assert_true(reader.result === null, "The value of result is " + reader.result);
                    t.done();
                });
            } else {
                t.step(function () {
                    assert_unreached("read state error: " + reader.readyState);
                });
                t.done();
            }
        } catch (ex) {
            t.step(function () {
                assert_unreached("throws exception: " + ex);
            });
            t.done();
        }
        window.setTimeout(function() {
            t.step(function () {
                assert_unreached("abort event is not fired");
            });
            t.done();
        }, 3000);
    </script>
  </body>
</html>
